requirements.txt
python-telegram-bot==21.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic-settings==2.5.2
python-dotenv==1.0.1
